Add files via upload
After building a classification model, there are various ways to evaluate the model. In the previous lessons, we've used classification report which includes precision, recall and f1-score on each classes, and confusion matrix to measure the performance of the classification models.

In this notebook, we will explain these evaluation technics in more detail.

We will demonstrate the classification metrics with adult income data from UCI data repository. We will also have 2 case studies on how to improve our model for different situations.